+++ src/com/intellij/idea/plugin/hybris/impex/psi/ImpexFile.kt
@@ -23,12 +23,18 @@ import com.intellij.idea.plugin.hybris.impex.ImpexLanguage
 import com.intellij.idea.plugin.hybris.impex.file.ImpexFileType
 import com.intellij.idea.plugin.hybris.impex.lang.folding.ImpexMacroDescriptor
 import com.intellij.idea.plugin.hybris.psi.util.getLineNumber
+import com.intellij.openapi.module.ModuleUtil
+import com.intellij.openapi.roots.ModuleRootManager
 import com.intellij.openapi.util.Key
+import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.psi.FileViewProvider
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiManager
 import com.intellij.psi.util.*
 import org.apache.commons.collections4.MultiValuedMap
 import org.apache.commons.collections4.multimap.HashSetValuedHashMap
+import org.apache.commons.lang3.StringUtils
+import java.io.File
 import java.io.Serial
 
 class ImpexFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, ImpexLanguage) {
@@ -54,6 +60,50 @@ class ImpexFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, Impe
         )
     }, false)
 
+    fun getExternalImpExFiles(): Collection<ImpexFile> = CachedValuesManager.getManager(project).getCachedValue(this, CACHE_KEY_EXTERNAL_FILES, {
+        val externalImpExFiles = mutableListOf<ImpexFile>()
+
+        this.acceptChildren(object : ImpexVisitor() {
+            override fun visitScript(o: ImpexScript) {
+                super.visitScript(o)
+
+                resolveIncludeExternalData(o)
+                    ?.let { externalImpExFiles.add(it) }
+            }
+        })
+
+        CachedValueProvider.Result.createSingleDependency(
+            externalImpExFiles,
+            PsiModificationTracker.MODIFICATION_COUNT,
+        )
+    }, false)
+
+
+    fun resolveIncludeExternalData(impexScript: ImpexScript): ImpexFile? {
+        val text = impexScript.text
+        val streamIndex = text.indexOf("impex.includeExternalData")
+            .takeIf { it != -1 }
+            ?.let { text.indexOf("getResourceAsStream") }
+            ?.takeIf { it != -1 }
+            ?: return null
+        val startIndex = text.indexOf('(', streamIndex).takeIf { it != -1 }
+            ?: return null
+        val endIndex = text.indexOf(')', startIndex).takeIf { it != -1 }
+            ?: return null
+
+        val resource = StringUtils.strip(text.substring(startIndex + 1, endIndex), "\"' ")
+        val module = ModuleUtil.findModuleForPsiElement(impexScript)
+            ?: return null
+
+        return ModuleRootManager.getInstance(module).sourceRoots
+            .filter { it.path.endsWith("/resources") }
+            .map { File(it.path, resource) }
+            .mapNotNull { LocalFileSystem.getInstance().findFileByIoFile(it) }
+            .firstOrNull { it.exists() }
+            ?.let { PsiManager.getInstance(impexScript.project).findFile(it) }
+            ?.let { it as? ImpexFile }
+    }
+
     fun getSuitableMacroDescriptor(macroName: String, macroUsage: PsiElement) = getMacroDescriptors()[macroName]
         ?.asSequence()
         ?.map { it.psiElement.getLineNumber() to it }
@@ -65,6 +115,7 @@ class ImpexFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, Impe
     companion object {
         val CACHE_KEY_HEADER_LINES = Key.create<CachedValue<Map<ImpexHeaderLine, Collection<ImpexValueLine>>>>("SAP_CX_IMPEX_HEADER_LINES")
         val CACHE_KEY_MACRO_DESCRIPTORS = Key.create<CachedValue<MultiValuedMap<String, ImpexMacroDescriptor>>>("SAP_CX_IMPEX_MACRO_DESCRIPTORS")
+        val CACHE_KEY_EXTERNAL_FILES = Key.create<CachedValue<Collection<ImpexFile>>>("SAP_CX_IMPEX_EXTERNAL_FILES")
 
         @Serial
         private val serialVersionUID: Long = 5112646813557523662L